chore: 어드민 유저 스키마 userSchema 파일로 이동
