+++ src/router-pages/admin/user/AdminMentorPage.tsx
@@ -1,3 +1,5 @@
+/** 멘토 관리 페이지 */
+
 export default function AdminMentorPage() {
   return <div>AdminMentorPage</div>;
 }
