+++ src/schema.ts
@@ -1516,34 +1516,6 @@ export const liveApplicationPriceType = z.object({
   livePriceType: livePriceTypeSchema,
 });
 
-/** GET /api/v1/user/admin */
-export const userAdminType = z.object({
-  userAdminList: z.array(
-    z.object({
-      userInfo: z.object({
-        id: z.number(),
-        name: z.string(),
-        email: z.string(),
-        contactEmail: z.string().nullable(),
-        phoneNum: z.string(),
-        createdDate: z.string(),
-        accountType: accountType.nullable(),
-        accountNum: z.string().nullable(),
-        marketingAgree: z.boolean().nullable(),
-      }),
-      applicationInfos: z.array(
-        z.object({
-          programId: z.number().nullable(),
-          programTitle: z.string(),
-        }),
-      ),
-    }),
-  ),
-  pageInfo,
-});
-
-export type UserAdmin = z.infer<typeof userAdminType>['userAdminList'];
-
 export const userAdminDetailType = z.object({
   userInfo: z.object({
     id: z.string(),
